requirements.txt
pandas
numpy
openpyxl
xlrd
XlsxWriter
python-dateutil
pyjanitor
matplotlib
seaborn
python-dotenv
Office365-REST-Python-Client
streamlit